Adding a library method for adios2pio-nm and removing global variables
