+++ tools/adios2pio-nm/adios2pio-nm-lib.h
@@ -1,0 +1,11 @@
+#ifndef _ADIOS2PIO_NM_LIB_H_ 
+#define _ADIOS2PIO_NM_LIB_H_ 
+
+#include <string>
+
+using namespace std;
+
+int ConvertBPToNC(string infilepath, string outfilename, string piotype);
+void usage(string prgname, int mpirank);
+
+#endif /* #ifndef _ADIOS2PIO_NM_LIB_H_ */
